split up ingress.go
